test(plan-runs): split coordinator.completion.test.ts at describe seam (warren-c157)
Move the 'PR recovery and plan completion' describe block (9 tests) plus
its file-local setupSucceededChildNoPr helper into a new sibling
coordinator.pr-recovery.test.ts. Keeps the 'completion phase' describe
(8 tests) in coordinator.completion.test.ts. Both files now sit under
the 500-line global limit; no test bodies changed (17 pass total).
